airflow/requirements.txt
snowflake-connector-python==3.6.0
dbt-snowflake==1.7.3
requests==2.31.0
pandas==2.0.3
